+++ dejafu-tests/lib/Examples/Philosophers.hs
@@ -11,7 +11,9 @@ import           Common
 tests :: [TestTree]
 tests =
   [ testDejafuWay way defaultMemType "deadlocks" deadlocksSometimes (philosophers 3)
-  , testDejafuWay way defaultMemType "loops"     abortsSometimes    (philosophers 3)
+  , let settings = set lshowAborts True (fromWayAndMemType way defaultMemType)
+    in testDejafuWithSettings settings "loops (with aborts present)" abortsSometimes (philosophers 3)
+  , expectFail $ testDejafuWay way defaultMemType "loops (with aborts hidden)" abortsSometimes (philosophers 3)
   ]
 
 -- | Shorter execution length bound