+++ dejafu-tests/lib/Integration/SingleThreaded.hs
@@ -4,10 +4,11 @@ module Integration.SingleThreaded where
 
 import           Control.Exception         (ArithException(..),
                                             ArrayException(..))
-import           Test.DejaFu               (Failure(..), gives, gives', isAbort,
-                                            isDeadlock, isIllegalDontCheck,
+import           Test.DejaFu               (Condition(..), gives, gives',
+                                            isAbort, isDeadlock,
                                             isUncaughtException)
 import           Test.DejaFu.Settings      (defaultLengthBound)
+import           Test.DejaFu.Types         (Error(..))
 
 import           Control.Concurrent.Classy
 import           Control.Monad             (replicateM_)
@@ -275,7 +276,7 @@ hacksTests = toTestList
     , djfu "Failures abort the whole computation" (alwaysFailsWith isDeadlock) $
         dontCheck Nothing $ takeMVar =<< newEmptyMVarInt
 
-    , djfu "Must be the very first thing" (alwaysFailsWith isIllegalDontCheck) $ do
+    , djfuE "Must be the very first thing" LateDontCheck $ do
         v <- newEmptyMVarInt
         dontCheck Nothing $ putMVar v 5